Move baaahs.Model et al. into baaahs.model.*.
