+++ src/commonMain/kotlin/baaahs/Show.kt
@@ -1,5 +1,7 @@
 package baaahs
 
+import baaahs.model.Model
+
 /** A show takes input from gadgets and uses it to configure shaders, creating pretty stuff on surfaces. */
 abstract class Show(val name: String) {
